requirements.txt
# Core dependencies
requests>=2.31.0
beautifulsoup4>=4.12.0
sentence-transformers>=2.2.2
faiss-cpu>=1.7.4
python-dotenv>=1.0.0
tqdm>=4.66.0
validators>=0.22.0

# Development dependencies
black>=23.12.0
flake8>=7.0.0
pylint>=3.0.3
mypy>=1.8.0
pytest>=7.4.3
pytest-cov>=4.1.0 